pyproject.toml
[tool.poetry]
name = "smit-test"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "smit_test"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.100.0"
tortoise-orm = "^0.19.3"
uvicorn = "^0.22.0"
pydantic-settings = "^2.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
